Refactor: move backup import view to a separate module
